code/admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Franco Giovanni Cirielli <ciriellifrancogiovanni@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
flask = "^3.0.3"
flask-session = "^0.8.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
flask-jwt-extended = "^4.6.0"
flask-bcrypt = "^1.0.1"
psycopg2-binary = "^2.9.9"
minio = "^7.2.9"
wtforms-sqlalchemy = "^0.4.1"
plotly = "^5.24.1"
pandas = "^2.2.3"
authlib = "^1.3.2"
flask-mail = "^0.10.0"
requests = "^2.32.3"
marshmallow = "^3.23.1"


[tool.poetry.group.dev.dependencies]
pytailwindcss = "0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
